Mark elm-review quick fixes as not available in batch mode.
If you apply an elm-review fix, you need to get fresh errors/fixes from elm-review for it to be up-to-date. The inline errors do this with the elm-review watch runner, but I can't figure out how to do this with a Global Inspection. So for now they are only for navigating to errors.